+++ api/src/models/types/request-bodies.ts
@@ -4,9 +4,13 @@ import {
 	AuthorizeSubscriptionBodySchema,
 	CreateChannelBodySchema,
 	RequestSubscriptionBodySchema
-} from '../schemas/request-bodies';
+} from '../schemas/channel-bodies';
+import { RevokeVerificationSchema, VerifyIdentitySchema } from '../schemas/verification-bodies';
 
 export type CreateChannelBody = Static<typeof CreateChannelBodySchema>;
 export type AddChannelLogBody = Static<typeof AddChannelLogBodySchema>;
 export type AuthorizeSubscriptionBody = Static<typeof AuthorizeSubscriptionBodySchema>;
 export type RequestSubscriptionBody = Static<typeof RequestSubscriptionBodySchema>;
+
+export type VerifyIdentityBody = Static<typeof VerifyIdentitySchema>;
+export type RevokeVerificationBody = Static<typeof RevokeVerificationSchema>;
