+++ api/src/models/types/verification.ts
@@ -1,0 +1,11 @@
+import Express from 'express';
+import { User } from './user';
+
+export interface AuthenticatedRequest extends Express.Request {
+	user: User;
+}
+
+export interface AuthorizationCheck {
+	isAuthorized: boolean;
+	error?: Error | null;
+}
