+++ api/src/routes/channel-info/index.ts
@@ -4,7 +4,7 @@ import { ChannelInfoService } from '../../services/channel-info-service';
 import * as _ from 'lodash';
 import { StatusCodes } from 'http-status-codes';
 import { getDateFromString } from '../../utils/date';
-import { AuthenticatedRequest } from '../../models/types/authentication';
+import { AuthenticatedRequest } from '../../models/types/verification';
 import { AuthorizationService } from '../../services/authorization-service';
 
 export class ChannelInfoRoutes {
