+++ api/src/routes/verification/index.ts
@@ -3,7 +3,8 @@ import { StatusCodes } from 'http-status-codes';
 import { VerifiableCredentialJson } from '../../models/types/identity';
 import { VerificationService } from '../../services/verification-service';
 import { Config } from '../../models/config';
-import { AuthenticatedRequest, AuthorizationCheck, RevokeVerificationBody, VerifyIdentityBody } from '../../models/types/authentication';
+import { RevokeVerificationBody, VerifyIdentityBody } from '../../models/types/request-bodies';
+import { AuthenticatedRequest, AuthorizationCheck } from '../../models/types/verification';
 import { UserService } from '../../services/user-service';
 import { User, UserRoles } from '../../models/types/user';
 import * as KeyCollectionLinksDb from '../../database/verifiable-credentials';
